The SMAnimation framework can also be used to describe very complicated, even nondeterministic, motions. This is done by declaring a SMAnimation with None for the images, and then passing it in to the at clause.
All motions (Motion, Pan, Move, and anim.SMAnimation when used as a
motion) can now be used as transitions, in which case the motion is
applied to the screen (or layer, as appropriate). This let us put
together two new transitions, hpunch and vpunch, which shake the
screen horizontally and vertically, as appropriate. Useful for when
the POV character gets punched by a particularly feisty girl.



git-svn-id: file:///home/tom/magnetic/ab/cvs-and-svn/svn-repos/renpy/trunk@134 ef037c1b-3fee-0310-a5a7-a1260de54e99