A new im.Tile image operator has been introduced. This operator tiles a source image until it reaches a specified size, or the size of the screen if no size is specified.
Ren'Py now supports a presplash screen. If the file presplash.png is
present in the game directory, it is shown to the user on
startup. This occurs before any loading or initialization occurs, so
the size of the window it is shown in is determined by the size of the
presplash.png image itself. 



git-svn-id: file:///home/tom/magnetic/ab/cvs-and-svn/svn-repos/renpy/trunk@166 ef037c1b-3fee-0310-a5a7-a1260de54e99