A new widget, anim.Blink, blinks a child widget in and out of existence by varying its alpha channel. [doc]
Character now allows for the use of a click-to-continue widget when
the widget finishes showing text to the user. While many widgets can
be used, this feature is especially useful with Images, Animations,
and anim.Blink. The click-to-continue widget can be nestled in with
the end of the text, or placed at a fixed location on the screen.



git-svn-id: file:///home/tom/magnetic/ab/cvs-and-svn/svn-repos/renpy/trunk@234 ef037c1b-3fee-0310-a5a7-a1260de54e99