The scene, show, and hide statements now take an onlayer clause, which gives the layer that the operations occur on. So if we declare high to be another layer (by setting config.layers to something like [ 'master', 'high', 'transient', 'overlay' ]), then one can use a statement like "show eileen happy onlayer high" to show the given image on the high layer. This image will not be removed when scene statements clear other layers, such as the default master layer.
To support the above, the parameters to renpy.show() changed in an
incompatible way. If your code uses renpy.show(), it will need to be
upgraded. Instead of considering the second and later parameters to be
at expressions, it now takes as its second parameter a list of all at
expressions. (This lets us use the third parameter to accept the layer
name.) 

The label 'after_load' is now called after a game is loaded, if it
exists on disk. This label can be used to upgrade save files, restore
music, and the like.

It's now possible to pick the screen the game menu enters when it is
invoked, by setting the _game_menu_screen variable.

Simplified the demo game script by moving the style stuff out of that
first init block, and down to the bottom of the script where it's out
of the way of a first viewing. Hopefully, this will reduce the shock
of someone first encountering Ren'Py.

Documented how to modify and replace the game and main menus, and how
to add in one's own preferences.



git-svn-id: file:///home/tom/magnetic/ab/cvs-and-svn/svn-repos/renpy/trunk@242 ef037c1b-3fee-0310-a5a7-a1260de54e99