This release introduces time correctness to the display subsystem of Ren'Py. Previously, time was a kinda flexible concept in Ren'Py. Objects were shown sometime before the first frame of an interaction, and the frames that we drew represented a smear of time, rather than some precise instant. This release changes that. Each frame now represents a single instant in time, and the timebases for things like animations, motions, and transitions all begin at the instant of the first frame of the first interaction in which they are present.
We also have precisely defined two timebases that displayables can
use. The displayable timebase starts at the instant of the first frame
in which a given displayable was present, while the animation timebase
begins at the instant of the first frame in which an image with the
same tag is present. Having two timebases makes it easier to specify
what we want happening when, say, an animation is placed inside a
motion. Animations default to the animation timebase, while motions
default to the motion timebase, but both have an option
(anim_timebase) that lets you specify which timebases is wanted. For
example, one may want to switch a looping motion onto the animation
timebase if one wants to allow the character's expression to change
while the motion continues.

We've implemented a more precise timeout system in Ren'Py. Previously,
if we wanted to have an event occur at some time after a displayable
has been shown, we'd have to use use the 20hz periodic event and deal
with it as part of that. The meant that our best timing precision was
50ms, and often it was significantly worse. Now, we can schedule
precise timeouts for event delivery, down to the ~10ms range. Among
other things, this allows for more accurate pauses, and skipping to
occur at a rate faster than the previous limit of 10 screens per
second or so.

Fixed a bug that cause Animations and anim.Filmstrips to ignore
position properties supplied to them. Now, these object properly
respect position properties supplied to them.



git-svn-id: file:///home/tom/magnetic/ab/cvs-and-svn/svn-repos/renpy/trunk@400 ef037c1b-3fee-0310-a5a7-a1260de54e99