Clarified that the voice extra can take non-wave sound files, such as OGG Vorbis and especially OGG Speex files. Also, made voice automatically sustain through pauses.
Fixed a bug in Ren'Py that was causing the voice extra to
malfunction. The bug was caused by the contents of the config
namespace being copied at the wrong time, leading it to not be
restored properly to default values when Ren'Py was reset (as the Main
Menu choice.) 

Removed a leftover debugging statement that caused Ren'Py to spew out
all of the say statements in .rpy files it parsed.