Rewrite MoveTransition, changing the API.
Movetransition now interpolates the positions of the image
in the before and after cases. This means that if the image is
moving in one or both of the cases, it will be moved between the
two locations.

The new movetransition uses transforms to specify where to move
to and from, rather than factory functions.

It's easier to use, but somewhat less flexible.

The old move transition is around as OldMoveTransition, and compat
will swap it in based on the script version.