Video sync and other video improvements.
We now generally synchronize video to wall time. We adjust this
synchronization slowly if we're slightly out of sync with the audio,
or by a lot if we're way out of sync.

This ensures that the frame rate remains stable even if the times of
audio decode vary a bit.

We also now support the Movie displayable again.