Add non-greedy rollback.
Greedy rollback rolls back as far as it can without crossing a
checkpoint. Non-greedy rollback rolls back to the start of
the checkpoint statement, or the start of the current
statement if we're in a statement.

(This is used by the console to get back to the start of the
current statement.)