Changes to persistence.
Automatically saving persistent data may be too slow (and might cause
a laptop's hard drive to spin up), so we now save it at shutdown or
when renpy.save_persistent is called.

We also allow creators to register a merge function that's called when
loading updated persistence data from disk.