Deal with fonts with odd bounding-boxes.
Some truetype fonts have bounding boxes that contain the origin. (This
may be becuase of font design, auto-hinthing, or anti-aliasing.) We
deal with these fonts by rendering all the available pixels, even
those to the left of the origin.

This may make the actual area drawn to the screen bigger than the
displayable itself, so we deal with that when necessary.

Fixes #34.