Stroke both font borders when creating outlines.
Previously we only stroked the outside border. This caused problems
when a portion of the character had an interior border on the
outside.

For example, we encountered a ? character where the top
part was outside and the bottom part was inside - it seems to
have something to do with the clockwise/counterclockwise order
of the font's control points.