Allow rpy files to fully support indentation with spaces and tabs.
This is 100% backwards compatible with all existing rpy files that use
spaces. Therefore, this does not require any change to anybody's
existing code. It only adds a new feature, allowing those who wish to
use tabs for indentation in rpy files to do so.

Previously, rpy files completely rejected indentation with tabs, and
forced all users to use spaces for indentation. Now, future rpy files
can be created using either tabs or spaces for indentation, at the
user's discretion.

To be sure that tabs and spaces are never mixed in a confusing manner,
this code forces all indentation to have the same prefix. Tabs are never
considered to be equivalent to any number of spaces. This matches what
we'd expect from files indented using best practices, with either spaces
only or tabs only, and eliminates any accidental "compatibility" between
the two in the same block.