Copy Fixeds when they're parameterized.
Parameterization (a terrible name) is used to reset transforms. In
cases where a single transform should be reset, a Fixed full of
tranforms should be replaced with a Fixed full of reset transforms.

(This is especially true for Fixeds created as part of a multiple
contains block in an ATL transform.)