SLDisplayables can't become constant if a child fails.
We don't want to mark a tree of displayables constant if a displayable
in that tree throws an error during prediction, since the tree will
be incomplete. (This is possible if the displayables in the tree
access missing non-parameter data, like a Text that interpolates a
missing variable.)

We propagate such fails up the tree of contexts, and do not allow
a SLDisplayable to become constant if a failure has occured.