Update current_screen when changing focus.
Previously, current_screen was not updated during focus operations.
This meant that in hovered and unhovered, actions like
SetScreenVariable could be applied to the wrong screen (basically,
whatever screen happened to be lying around after the event
processing or update cycle).

After this change, we track the current screen as we change focus,
and update current_screen appropriately.

Fixes #610.