Ensure screens are prepared once on start.
Screen preparation is expensive, and needs to be done every time
styles are rebuilt. Previously, we had been rebuilding multiple
times in the start process. Now, on a normal start, we only
prepare once.

This required quite a bit of reorganization of init code, but
it should be worth it.