Improve default statement.
The default statement now runs after all rollbacks, not just loads.

It also stores the names of the default variables that have already
been set - this will allow a default to override a builtin present
in a future Ren'Py version.

Lastly, it checks for multiple defaults for a single variable, and
errors out if it finds them.