Add the init label statement.
This defines a label in an init parser context, so that
the style/image/define/etc statements inside will only
be run if the label is called.

This is intended to be part of the new theme system, allowing
a theme to be defined using style statements. It's not intended
for creator use.