Render-to-texture at the drawable resolution.
When an image is zoomed down and the drawable screen is bigger than
the virtual screen, Ren'Py will draw the additional information
present in the image. This information was being lost during RTT
operations that occured at the virtual resolution. (This usually
manifested as a blurry dissolve.)

Now, RTT occurs at the drawable resolution, and so the information
is shown to the player.

This also adds the information needed to render text at drawable
resolution.