Wrap shown transforms in an ImageReference.
Fixes #695.

The problem is that Ren'Py had no way of telling Transforms created by
an at-list from those supplied to renpy.show as input. The result is
that we'd get a stack of transforms, and then change the child the
entire stack, which could cause repeats in the transform stack
(see the bug).

This is fixed by wrapping Transforms passed to renpy.show in an
ImageReference, which makes it clear which are on the "outside"
and which are on the "inside" of the at_list.