Include possibly-focusable displayables in the focus order.
A possibly-focusable thing is a button/bar/etc that is currently
insensitive, but might change to become sensitive in the future.
We want to have those things reserve their place in the focus
order, so the thing with focus doesn't lose focus when
something else becomes focusable.

Fixes #725, maybe. This is my best guess as to what's going on
there.