Use virtual text size to compute screen placement.
A problem with the new text code is that the text can change its
virtual size. This can cause other layouts to jump around as the
screen is scaled.

To fix this, this change renders text twice, first at the virtual
size, and then at the drawable size. The first pass is used to
compute line breaking, and to get the size of the final Render.
Once that's done, the drawable size is used to draw cleaner text.

Fixes #728.