Better compute subsurface when blits go outside a Render.
Text drawn using the new rendering code can be drawn outside the
Render, due to the difference between virtual- and drawable-resolution
text layout. Subsurface would not crop to the intersection of the
crop-box and the render, which cut text off.

This changes to only crop to the crop-rect, which means that text
leaving the Render ist still drawn (as long as it's in the crop-rect).

See: http://lemmasoft.renai.us/forums/viewtopic.php?f=8&t=31311&start=120#p385132