text: Adjust spacing of drawable-res text to fill virtual-res box.
It's possible for drawable-res text to be bigger or smaller than
the virtual-res box. This can lead to ugliness in text layout, as
centered text will not be centered, margins will be uneven, etc.

This fix adds code that adjust glyph placement to ensure the
drawable glyphs fill the same place as the physical glyphs,
using techniques similar to kerning and line_spacing.