ffmedia: Synchronize video display and frame dropping. [doc]
When multiple videos are playing, we need them to play in lockstep,
to ensure that frames stay in sync. (This is especially important
for video sprites, to ensure the color and alpha images have the
same pts.)

When multiple movies are playing, Ren'Py now checks that there is a
frame available from every playing movie before reading the frame from
any movie. As long as the movies have the same framerate, they should
always remain in lockstep.

Framedropping now works by dropping all frames that should be
displayed before the time the last frame was actually displayed.
This ensures that framedropping is consistent on all playing
movies. (Again, assuming the framerates are the same.)