Support pre-6.11 games that overwrite config.layers.
Prior to 6.11, screens weren't a thing and as such there wasn't a "screens" layer.
If games before this time unwisely overwrote the entirety of config.layers,
the screens layer would go missing, leading to all kinds of nasty and unexpected
behaviour later on. An example of a game that can be ran on modern Ren'Py
with this fix is Digital: A Love Story by Christine Love.