Defer styles and change the order of translate style blocks.
The mofivation behind this is that we want translate code to be
able to use a translate python statement to update a couple of
variables that control fonts and sizes, rather than having to
update every style that uses the font or size.

To make this work, we add a way of deferring style statements
until after the translate python code has run, and change
the order of translate python and style code.