Change Displayable._get_parameterized to ._target.
This is used to get the ultimate target of one or more
ImageReferences, especially if that target is a displayable.

This used to be a lot more complicated, but since going forwards,
all ImageReferences will be parameterized, it makes sense to
use a simpler name and semantics for the method.