gui: Add space to the left of the check/radio bar image.
This makes the result more visually pleasing, as the image
lines up properly with vertical lines in DejaVuSans glyphs.
(For example, the left side of D in display.) This is offset to the
right of the text bounding box by ~3 pixels, probably so slanted
lines can go a little to the left of it.