Change the name of 6.99.11.
Needed to get in a Hamilton reference somewhere, since I've been
listening to it Non-Stop while making this version. And that seems
to be another reference.

(And I'll mention Silver here, too - the car we took on our road
trip to Dallas and Hot Springs. A 2016 Dodge Challenger rental,
VIN 2C3CDZAG1GH254389, we had her from miles 2302-4557, 2255 miles
total. We listened to Hamilton for probably over 150 of those
miles...)