Change how we build and version updates.
This fixes a problem where, since current.json was not included,
the DLC was not being updated properly, meaning that updates
would also be required.

Fixing this (while still preserving fast building) required changing
the format of upgrades a little. The pretty version number is no
longer included in update files (it wasn't used), and the versions
themselves are hashes of the file content. This allows us to include
current.json in update files without having to constantly regenerate
those files.