Preserve looping across renpy.music.play(..., if_changed=True).
This fixes a problem where renpy.music.play(..., if_changed=True)
would requeue the current track, which would cause it to start
playing from the beginning, rather than the loop point.

Now, if the music hasn't changed we just update the loop, which
causes it to start playing from the loop point.

Fixes #1055.