Change the default encoding of python to the fsencoding.
This is - loosely - to improve compatibility with Python 3 and
the changes introduced in 6.99.12. In Python 3, all strings
entering from the OS will be unicode - and hence, unicode can
be the filesystem encoding.

Before this change, strings from the OS were considered ASCII,
and hence would fail when concatenated to the unicode. Now they're
considered to be in the filesystem encoding, which Ren'Py already
handles properly.

This should work since all the strings in Ren'Py should be in either
unicode, utf-8 (converted explicity to or from unicode), the
filesystem encoding, ASCII (in all relevant codecs, compatible with
the filesystem encoding), or some binary format that's not converted
to unicode.