Pop dynamic data from non-root contexts while saving.
Fixes #1160.

The problem here is that renpy.retain_after_load will prevent the
rollback that usually occurs on loading - and that rollback was
responsible for getting rid of all of the dynamic data that
is used in the menu context. This code pops the dynamic data
in a more explicit way.