Improve how clipping and subsurface work.
There are two improvements here. The first is that it's now
possible for a render to specify clipping in the x axis, y axis,
or both.

The second is that when taking a subsurface, Ren'Py will not clip
surfaces that fit entirely inside the clipping rectange.

This should completely fix the cut-off text issue from #1113.