Keep track of the texture generation by texture number.
It seems like resizing the screen can reset glGenTexures, leading
to the scenario where a texture number is allocated twice. By
keeping a map from texture number to texture, we can check
to make sure the generation associated with the texture
matches that associated with the texture number, and prevent
double-deallocation problems.

Fixes #1246.