NPOT fix and improvements.
NPOT was rewritten to accept the maximum size texture to use
from the rest of Ren'Py. This ensures that we do not go past
the edge of the screen when doing a copy RTT, something that
causes the texture to be cut off.

As part of this allowed Ren'Py to do a screen-sized RTT
(as is done for Dissolve) in a single operation, by
removing the 1px border that had been present
around textures. This required us to enable GL_CLAMP_TO_EDGE
to effectively provide these borders.

Fixes #1249.