Expand the clip rectangle by a small amount.
This prevents a numerical stability issue. The clipping rectangle
and the draw rectangle can be the same, but computed in different
ways. This means the clip test can reject a pixel that should be
drawn. Very small offsets prevent this case.

This fixes the symptom of 1px lines appearing in tiled frames.