Only cycle the storedicts on log.begin.
Previously, the storedicts were being cycled on log.complete(),
which is called during a save as well as normal statement begin.
When cycled, the storedicts would record the information while
in the save context, and restore it.

This was causing problems with suppress_overlay, and less likely
other problems.

Fixes #1383.