Separate the screen cache into generations.
This fixes a problem where the cache of a screen could contain
a transform in the old state. What's more, since the cache was
persisting between screens, showing two screens with the same
tag would use the (long obsolete) cached data, and hide the
old screen.

By breaking things up into old and new generations, a transform
not used in the new generation will "go away", and be reset when
shown again.